fix: add validation when category deleted

+++ src/main/kotlin/dev/trodrigues/dscatalogapi/services/impl/CategoryServiceImpl.kt
@@ -3,8 +3,10 @@ package dev.trodrigues.dscatalogapi.services.impl
 import dev.trodrigues.dscatalogapi.domain.Category
 import dev.trodrigues.dscatalogapi.extension.toModel
 import dev.trodrigues.dscatalogapi.repositories.CategoryRepository
+import dev.trodrigues.dscatalogapi.repositories.ProductRepository
 import dev.trodrigues.dscatalogapi.resources.requests.PutCategoryRequest
 import dev.trodrigues.dscatalogapi.services.CategoryService
+import dev.trodrigues.dscatalogapi.services.exceptions.DomainException
 import dev.trodrigues.dscatalogapi.services.exceptions.ObjectNotFoundException
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
@@ -13,7 +15,8 @@ import org.springframework.transaction.annotation.Transactional
 
 @Service
 class CategoryServiceImpl(
-    private val categoryRepository: CategoryRepository
+    private val categoryRepository: CategoryRepository,
+    private val productRepository: ProductRepository
 ) : CategoryService {
 
     override fun findAll(pageable: Pageable): Page<Category> = categoryRepository.findAll(pageable)
@@ -36,6 +39,9 @@ class CategoryServiceImpl(
     @Transactional
     override fun delete(categoryId: Long) {
         val category = findById(categoryId)
+        if(productRepository.existsByCategories(category)) {
+            throw DomainException("Category cannot be deleted, category($categoryId) has product registered")
+        }
         categoryRepository.delete(category)
     }
